# Refactor /projects/:id API handler