--- a/carbonmark-api/helpers/fetchProjectDetails/fetchGSProject.js
+++ b/carbonmark-api/helpers/fetchProjectDetails/fetchGSProject.js
@@ -1,0 +1,8 @@
+export const fetchGSProject = async ({ registryProjectId }) => {
+  //   const res = await fetch(`https://api.goldstandard.org/projects/${id[1]}`);
+  //   const data = await results.json();
+  //   return {
+  // description: data.description,
+  //   };
+  return null; // gs not currently supported
+};