--- a/carbonmark-api/helpers/fetchProjectPoolInfo.js
+++ b/carbonmark-api/helpers/fetchProjectPoolInfo.js
@@ -1,7 +1,7 @@
-import { poolInfo } from "../constants/constants";
-const { executeGraphQLQuery } = require("../apollo-client");
-const { POOL_PROJECTS } = require("../queries/pool_project");
 // @ts-check
+import { executeGraphQLQuery } from "../apollo-client";
+import { poolInfo } from "../constants/constants";
+import { POOL_PROJECTS } from "../queries/pool_project";
 
 /**
  * @typedef {"bct" | "nct" | "ubo" | "nbo"} PoolName
@@ -35,32 +35,61 @@ const { POOL_PROJECTS } = require("../queries/pool_project");
  * @property {string} balanceNCT
  * @property {string} balanceBCT
  * @property {string} tokenAddress
+ * @property {string} totalBridged
+ * @property {string} totalRetired
+ * @property {string} currentSupply
+ */
+
+/**
+ * @typedef {Object} Stats
+ *   Stats for all project tokens across both bridges
+ * @property {number} totalBridged
+ * @property {number} totalRetired
+ * @property {number} totalSupply
  */
 
 /**
  * @typedef {Object} Params
  * @property {string} key - Project key `"VCS-981"`
- * @property {string} vintageStr - Vintage string `"2017"`
+ * @property {string} vintage - Vintage string `"2017"`
  */
 
 /**
  * Query the subgraph for a list of the C3Ts and TCO2s that exist for a given project-vintage combination.
  * @param {Params} params
- *  @example fetchCarbonProjectTokens({ key: "VCS-981", vintageStr: "2017" })
- * @returns {Promise<PoolInfoMap>}
+ *  @example fetchCarbonProjectTokens({ key: "VCS-981", vintage: "2017" })
+ * @returns {Promise<[PoolInfoMap, Stats]>}
  * A map of project token info for each pool. For example VCS-981-2017 has been bridged to a C3T (pooled in NBO) and a TCO2 (pooled in NCT)
  */
 export const fetchProjectPoolInfo = async (params) => {
   const { data } = await executeGraphQLQuery(
     process.env.CARBON_OFFSETS_GRAPH_API_URL, // polygon-bridged-carbon subgraph
     POOL_PROJECTS,
-    { key: params.key, vintageStr: params.vintageStr }
+    { key: params.key, vintageStr: params.vintage }
   );
 
   /** @type {QueryResponse[]} */
   const tokens = data?.carbonOffsets?.length ? data.carbonOffsets : [];
 
-  return Object.keys(poolInfo).reduce((prevMap, poolName) => {
+  /** @type {Stats} */
+  const initialStats = {
+    totalBridged: 0,
+    totalRetired: 0,
+    totalSupply: 0,
+  };
+
+  const stats = tokens.reduce((stat, token) => {
+    return {
+      totalBridged: stat.totalBridged + Number(token.totalBridged),
+      totalRetired: stat.totalRetired + Number(token.totalRetired),
+      totalSupply: stat.totalSupply + Number(token.currentSupply),
+    };
+  }, initialStats);
+
+  /** @type {any} */
+  const initialMap = {};
+
+  const poolInfoMap = Object.keys(poolInfo).reduce((prevMap, poolName) => {
     const balanceKey = `balance${poolName.toUpperCase()}`; // e.g. "balanceBCT"
     // see if any of the entries from the subgraph have a balance > 1, otherwise return "0"
     const matchingTokenInfo = tokens.find((t) => Number(t[balanceKey]) > 1);
@@ -71,13 +100,15 @@ export const fetchProjectPoolInfo = async (params) => {
         supply: matchingTokenInfo ? matchingTokenInfo[balanceKey] : "0",
         poolAddress: poolInfo[poolName].poolAddress,
         isPoolDefault: matchingTokenInfo
-          ? matchingTokenInfo.tokenAddress ===
-            poolInfo[poolName].defaultProjectTokenAddress
-          : "",
+          ? matchingTokenInfo.tokenAddress.toLowerCase() ===
+            poolInfo[poolName].defaultProjectTokenAddress.toLowerCase()
+          : false,
         projectTokenAddress: matchingTokenInfo
           ? matchingTokenInfo.tokenAddress
           : "",
       },
     };
-  }, {});
+  }, initialMap);
+
+  return [poolInfoMap, stats];
 };