--- a/carbonmark-api/queries/all_pool_prices.js
+++ b/carbonmark-api/queries/all_pool_prices.js
@@ -1,0 +1,12 @@
+const { gql } = require("@apollo/client");
+
+const ALL_POOL_PRICES = gql`
+  query pairs {
+    prices: pairs {
+      address: id
+      price: currentprice
+    }
+  }
+`;
+
+module.exports.ALL_POOL_PRICES = ALL_POOL_PRICES;