… carbonmark-api/routes/users.js
… carbonmark/components/Activities/Activity.tsx
… carbonmark/components/ProjectCard/index.tsx
… carbonmark/components/Stats/StatsBar.tsx
… carbonmark/components/Stats/index.tsx
… carbonmark/components/pages/Home/index.tsx
… carbonmark/components/pages/Project/BuyOptions/PoolPrice.tsx
… carbonmark/components/pages/Project/BuyOptions/SellerListing.tsx
… carbonmark/components/pages/Project/Purchase/Pool/AssetDetails.tsx
… carbonmark/components/pages/Project/Purchase/Pool/PurchaseForm.tsx
… carbonmark/components/pages/Project/Purchase/Pool/PurchaseInputs.tsx
… carbonmark/components/pages/Project/Purchase/Pool/SuccessScreen.tsx
… carbonmark/components/pages/Project/Purchase/Pool/TotalValues.tsx
… carbonmark/components/pages/Project/Purchase/index.tsx
… carbonmark/components/pages/Project/Retire/AssetDetails.tsx
… carbonmark/components/pages/Project/Retire/RetireForm.tsx
… carbonmark/components/pages/Project/Retire/RetireInputs.tsx
… carbonmark/components/pages/Project/Retire/TotalValues.tsx
… carbonmark/components/pages/Project/index.tsx
… carbonmark/components/pages/Purchases/index.tsx
… carbonmark/components/pages/Retire/index.tsx
… carbonmark/lib/constants.ts
… carbonmark/lib/createUrls.ts
… carbonmark/lib/listingsGetter.ts
… carbonmark/lib/types/carbonmark.ts
… carbonmark/pages/projects/[project_id]/purchase/pools/[pool].ts
… carbonmark/pages/projects/[project_id]/retire/pools/[pool].ts
… examples/nextjs-typescript-integration/app/[project_key]/page.tsx
… examples/nextjs-typescript-integration/components/ProjectInfoCard.tsx
… examples/nextjs-typescript-integration/utils/fetchProjectInfo.ts