## Description

I could not figure out a clean way to update the pricing to reflect default projects in the current code. So I refactored the entire handler.

- reduces the total number of roundtrip network requests significantly:
  - instead of 4 graphql queries (one for each pool), we do 1 query total (`fetchAllPoolPrices`)
  - removes the c3token and tco2token requests, these are not needed
  - parallelizes the firebase requests
- eliminates / consolidates a lot of code into utils
- move to a single-source-of-truth object/map for pool info

## Breaking changes:
- changed some api property names.
  - prices array more clear property names:
    - `name` -> `poolName`
    - `leftToSell` -> `supply`
    - `tokenAddress` -> `poolAddress`
    - `poolTokenAddress` -> `projectTokenAddress`
  - moved `currentSupply` `totalBridged` and `totalRetired` into a field called `stats`
  
## Performance

Performance has only improved slightly. We need to dig deeper into which requests are taking the longest.
Occasionally production would give a 3-6 second load time, and this does not seem to happen as often after this PR. But I can't reliably reproduce this.

### Before PR: Production (api.carbonmark.com)
```
// VCS-191-2008
getCarbonmarkProject: 919.481ms
// VCS-1883-2018
getCarbonmarkProject: 858.302ms
// VCS-1140-2015
getCarbonmarkProject: 1.142s
```

This pr (using preview link).
Time for initial fetch only:
```
// VCS-191-2008
getCarbonmarkProject: 677.088ms
// VCS-1883-2018
getCarbonmarkProject: 783.747ms
// VCS-1140-2015
getCarbonmarkProject: 609.366ms
```
  
## Related Ticket

Resolves #1156